+++ chain/state/cache/factory.go
@@ -1,4 +1,4 @@
-package object
+package cache
 
 import (
 	"context"
