+++ chain/state/cache/persistent_cache.go
@@ -1,4 +1,4 @@
-package object
+package cache
 
 import (
 	"context"
