+++ chain/state/cache/types.go
@@ -1,4 +1,4 @@
-package object
+package cache
 
 import (
 	"github.com/ethereum/go-ethereum/common"
